requirements.txt
fastapi
uvicorn
jinja2
python-multipart
gitpython
supabase
aiofiles
websockets
python-dotenv